refactor: 채팅, 게임방 관련 함수들 서비스 분리
